Only warn on private API notes having the wrong case when appropriate (#232)
Specifically, if the private API notes are for a top-level private
module named "FooKit_Private", that's different from being for a
public module named "FooKit". Don't warn in that case. (And rather
than check for the presence of a "FooKit_Private", look at the
submodules of "FooKit" and see if any of them come from a private
module map.)

rdar://problem/45705724