[Modules] Move search paths outside of the control block
Clang does not currently hashes the search paths, so it doesnt make sense
for them to be in the control block, and we should move those to the
UNHASHED_CONTROL_BLOCK instead.

One concern mentioned with moving the header search paths to the
unhashed control block is that we stop rebuilding modules in the non-PCH
scenario, which might lead to problems if the header search path
actually changes something we rely for consistency. However, the control
block already includes the INPUT_FILE list, which should also force the
rebuild in non-PCH builds, so this should be fine.

There are no testcases for this change, as it would require using
llvm-bcanalyzer in clang tests, which we don't currently do.

rdar://problem/45567811