[test] Set -fno-split-cold-code in tests sensitive to splitting, NFC
Set -fno-split-cold-code in a few clang tests which check the output of
a full optimization pipeline.

These tests are sensitive to changes in the pipeline in general, and to
a change which enables splitting by default in particular.

(cherry picked from commit b65498798e5f5c003300b357a509f57af18a62bb)