[ASTImporter] Handle built-in when importing SourceLocation and FileID
Summary:
Currently when we see a built-in we try and import the include location. Instead what we do now is find the buffer like we do for the invalid case and copy that over to the to context.

Differential Revision: https://reviews.llvm.org/D58743

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@355332 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 2f3e65f35906044e5f7dcd39bcd119e63db29fc7)