Non-fragile ObjC ABI: Add support for __attribute__((objc_class_stub))
This attribute denotes a class with resiliently-sized metadata, which
is initialized at runtime. Since there is no statically-emitted metadata
in this case, Clang must reference a "class stub" instead.

When emitting a classref load we have to call `objc_loadClassRef()`
instead of loading directly.

In the classref itself, the least significant bit of the class pointer
is set to 1.

The Objective-C runtime implements objc_loadClassRef() by checking if
the least significant bit of the class pointer is 1; if not, it simply
returns the class pointer. Otherwise, it calls an initialization hook
stored inside the class stub, and stores the result of the hook back
to the classref to fast path future loads.

Note that categories continue to be emitted as before; the runtime will
lazily attach categories when class stubs are realized.