[Modules] Make ReadModuleMapFileBlock errors reliable
During implicit module builds, there are cases where ReadASTCore is
called with ImportedBy set to nullptr, which breaks expectations in
ReadModuleMapFileBlock and lead to crashes instead of an error message.
Fix this by improving ReadModuleMapFileBlock to handle ImportedBy
properly.

Could not reproduce the exact scenario where this happens but looking at
a crash backtrace we could verify the ReadASTCore callsite sets
ImportedBy to nullptr.

There are no testcases, since we were not even able to reproduce the
issue, but there were lots of crashes with the same backtrace.

rdar://problem/48828801