[DependencyCollector] Add optional reproducer callback
The clang importer in Swift uses its own subclass of the clang
dependency collector. Because it is completely hidden in the clang
importer's implementation, we cannot subclass it from LLDB.

This is an issue for reproducers, because LLDB needs to be informed of
all the files used by clang through the Swift compiler. To solve this, I
added the ability to register a custom callback in the dependency
collector. LLDB can use this callback to be notified, without needing
access to the dependency collector's implementation.