Change while to do-while
The assert statement says that the location must be a macroID, which is true. However, the while statements checks for that again, which is unnecessary, and gives warnings for returning a potentially non-initialized variable. For this reason, I suggest changing the while to a do-while loop while keeping the assert where it is.